feat: improve authorization, tests, and error handling

+++ Backend/SPC.Business/Services/AuthorService.cs
@@ -51,12 +51,12 @@ public class AuthorService : IAuthorService
 
     public async Task<BaseMessage<Author>> FindById(int id)
     {
-        Author? album = new();
-        album = await _unitOfWork.AuthorRepository.FindAsync(id);
+        Author? author = new();
+        author = await _unitOfWork.AuthorRepository.FindAsync(id);
 
-        return album != null ?
-            BuildResponse(new List<Author>() { album }, "", HttpStatusCode.OK, 1) :
-            BuildResponse(new List<Author>(), "", HttpStatusCode.NotFound, 0);
+        return author != null ?
+            BuildResponse(new List<Author>() { author }, "", HttpStatusCode.OK, 1) :
+            BuildResponse(new List<Author>(), "Author Not Found", HttpStatusCode.NotFound, 0);
     }
 
     public async Task<BaseMessage<Author>> FindByName(string name)
@@ -140,6 +140,16 @@ public class AuthorService : IAuthorService
         try
         {
             author = await _unitOfWork.AuthorRepository.FindAsync(id);
+            if (author == null)
+            {
+                return new BaseMessage<Author>()
+                {
+                    Message = "Author Not Found",
+                    StatusCode = HttpStatusCode.NotFound,
+                    TotalElements = 0,
+                    ResponseElements = new()
+                };
+            }
             await _unitOfWork.AuthorRepository.Delete(id);
             await _unitOfWork.SaveAsync();
         }
@@ -161,6 +171,52 @@ public class AuthorService : IAuthorService
             TotalElements = 1,
             ResponseElements = new List<Author> { author }
         };
+    }
+
+    public async Task<BaseMessage<Author>> AddAuthors(List<Author> authors)
+    {
+        if (authors == null || !authors.Any())
+        {
+            return BuildResponse(new(), "The author list cannot be empty.", HttpStatusCode.BadRequest);
+        }
+
+        foreach (var author in authors)
+        {
+            var isValid = ValidateModel(author);
+            if (!string.IsNullOrEmpty(isValid))
+            {
+                return BuildResponse(new(), isValid, HttpStatusCode.BadRequest);
+            }
+        }
+
+        try
+        {
+            foreach (var author in authors)
+            {
+                await _unitOfWork.AuthorRepository.AddAsync(author);
+            }
+            await _unitOfWork.SaveAsync();
+        }
+        catch (Exception ex)
+        {
+            return new BaseMessage<Author>()
+            {
+                Message = $"[Exception]: {ex.Message}",
+                StatusCode = HttpStatusCode.InternalServerError,
+                TotalElements = 0,
+                ResponseElements = new()
+            };
+        }
+
+
+        return new BaseMessage<Author>()
+        {
+            Message = "",
+            StatusCode = HttpStatusCode.OK,
+            TotalElements = 1,
+            ResponseElements = authors.ToList()
+        };
+
     }
 
     private BaseMessage<Author> BuildResponse(List<Author> lista, string message = "", HttpStatusCode status = HttpStatusCode.OK, int totalElements = 0)
