+++ Backend/SPC.Business/Services/BookService.cs
@@ -62,7 +62,7 @@ public class BookService : IBookService
 
         return book != null ?
           BuildResponse(new List<Book>() { book }, "", HttpStatusCode.OK, 1) :
-          BuildResponse(new List<Book>(), "", HttpStatusCode.NotFound, 0);
+          BuildResponse(new List<Book>(), "Book Not Found", HttpStatusCode.NotFound, 0);
     }
 
     public async Task<BaseMessage<Book>> UpdateBook(Book book)
@@ -125,6 +125,16 @@ public class BookService : IBookService
         try
         {
             book = await _unitOfWork.BookRepository.FindAsync(id);
+            if (book == null)
+            {
+                return new BaseMessage<Book>()
+                {
+                    Message = "Book Not Found",
+                    StatusCode = HttpStatusCode.NotFound,
+                    TotalElements = 0,
+                    ResponseElements = new()
+                };
+            }
             await _unitOfWork.BookRepository.Delete(book);
             await _unitOfWork.SaveAsync();
         }
@@ -189,6 +199,52 @@ public class BookService : IBookService
         return await UpdateBook(book);
     }
 
+     public async Task<BaseMessage<Book>> AddBooks(List<Book> books)
+    {
+        if (books == null || !books.Any())
+        {
+            return BuildResponse(new(), "The book list cannot be empty.", HttpStatusCode.BadRequest);
+        }
+
+        foreach (var book in books)
+        {
+            var isValid = ValidateModel(book);
+            if (!string.IsNullOrEmpty(isValid))
+            {
+                return BuildResponse(new(), isValid, HttpStatusCode.BadRequest);
+            }
+        }
+
+        try
+        {
+            foreach (var book in books)
+            {
+                await _unitOfWork.BookRepository.AddAsync(book);
+            }
+            await _unitOfWork.SaveAsync();
+        }
+        catch (Exception ex)
+        {
+            return new BaseMessage<Book>()
+            {
+                Message = $"[Exception]: {ex.Message}",
+                StatusCode = HttpStatusCode.InternalServerError,
+                TotalElements = 0,
+                ResponseElements = new()
+            };
+        }
+
+
+        return new BaseMessage<Book>()
+        {
+            Message = "",
+            StatusCode = HttpStatusCode.OK,
+            TotalElements = 1,
+            ResponseElements = books.ToList()
+        };
+
+    }
+
     private BaseMessage<Book> BuildResponse(List<Book> lista, string message = "", HttpStatusCode status = HttpStatusCode.OK, int totalElements = 0)
     {
         return new BaseMessage<Book>()
@@ -228,9 +284,13 @@ public class BookService : IBookService
         {
             message += "El nombre es requerido";
         }
-        if (string.IsNullOrEmpty(book.ISBN13) || !ISBNIsValid(book.ISBN13))
+        if(string.IsNullOrEmpty(book.ISBN13))
         {
-            message += "El ISBN es requerido y con un formato válido";
+            message += "El ISBN13 es requerido";
+        }
+        if (!ISBNIsValid(book.ISBN13))
+        {
+            message += $"El ISBN debe tener un formato válido";
         }
         if (book.YearOfPubliction < 1450 || book.YearOfPubliction > DateTime.Now.Year)
         {
@@ -247,7 +307,8 @@ public class BookService : IBookService
 
     static bool ISBNIsValid(string input)
     {
-        return Regex.IsMatch(input, @"^[0-9-]+$");
+        return Regex.IsMatch(input, @"^978-\d{1,5}-\d{1,7}-?\d{0,7}-?\d?$");
+
     }
 
     #region Learning to Test
